% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_kmeans_to_dendrogram.R
\name{convert_kmeans_to_dendrogram}
\alias{convert_kmeans_to_dendrogram}
\title{converts a list of integer labels to a pseudo-tree for plotting above or along side a heat map
Authors:  Richard Moffitt, Steven Botez, Peter Ly}
\usage{
convert_kmeans_to_dendrogram(sample_order)
}
\description{
converts a list of integer labels to a pseudo-tree for plotting above or along side a heat map
Authors:  Richard Moffitt, Steven Botez, Peter Ly
}
