% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_order_to_dendrogram.R
\name{convert_order_to_dendrogram}
\alias{convert_order_to_dendrogram}
\title{create a flat dendrogram based on a ordered list of integers}
\usage{
convert_order_to_dendrogram(sampleorder)
}
\description{
create a flat dendrogram based on a ordered list of integers
}
