% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_SummarizedExperiment.R
\name{convert_to_SummarizedExperiment}
\alias{convert_to_SummarizedExperiment}
\title{Create SummarizedExperiment object from pdac list style object}
\usage{
convert_to_SummarizedExperiment(dataset)
}
\description{
Create SummarizedExperiment object from pdac list style object
}
