% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.classif.R
\name{create.classif}
\alias{create.classif}
\title{classify a data set with a trained TSP classifier}
\usage{
create.classif(dat, classifier, dec = NULL, drop = F, labels = NULL, fit)
}
\description{
classify a data set with a trained TSP classifier
}
