% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSideColors.R
\name{getSideColors}
\alias{getSideColors}
\title{create a sidecolor object from a sampInfo and a list of tracks}
\usage{
getSideColors(
  sampInfo,
  sampleTracks = names(sampInfo),
  colorlists = c("gray94", "blue", "green", "yellow", "orange", "red", "black"),
  displaynames = as.list(sampleTracks),
  drop.levels = FALSE
)
}
\description{
create a sidecolor object from a sampInfo and a list of tracks
}
