% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICGC_cluster_based_subtyping.R
\name{ICGC_cluster_based_subtyping}
\alias{ICGC_cluster_based_subtyping}
\title{Cluster Based Subtyping}
\usage{
ICGC_cluster_based_subtyping(data_set_list = c("summary"))
}
\description{
Perform clustering, label samples, and re-save data sets
}
