% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_cluster_based_subtyping.R
\name{other_cluster_based_subtyping}
\alias{other_cluster_based_subtyping}
\title{Cluster Based Subtyping}
\usage{
other_cluster_based_subtyping(data_set_list = c("summary"))
}
\description{
Perform clustering, label samples, and re-save data sets
}
