% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_moffitt_classifier.R
\name{parse_moffitt_classifier}
\alias{parse_moffitt_classifier}
\title{Parse moffitt tumor classifier from 2015 virtual microdissection paper}
\usage{
parse_moffitt_classifier()
}
\description{
Parse moffitt tumor classifier from 2015 virtual microdissection paper
}
