% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_scOh_atlas.R
\name{parse_scOh_atlas}
\alias{parse_scOh_atlas}
\title{This code is here for transparency, but for data storage reasons the full combined rds file is not available.}
\usage{
parse_scOh_atlas()
}
\description{
This code is here for transparency, but for data storage reasons the full combined rds file is not available.
}
